# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "老年人用药管理系统",
    "description": "老人用药管理智能助手后端 API",
    "version": "2.33.0"
  },
  "paths": {
...
    "/api/v1/public/device/status/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Status",
        "description": "获取设备状态信息（供子女端查询，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_status_api_v1_public_device_status__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/device/medication_plan": {
      "post": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Set Device Medication Plan",
        "description": "家属通过设备ID设置用药计划（校验 device_id 与 X-Device-Token）",
        "operationId": "set_device_medication_plan_api_v1_public_device_medication_plan_post",
        "parameters": [
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMedicationPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/plans/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Plans",
        "description": "获取设备的所有用药计划（供子女端查看，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_plans_api_v1_public_device_plans__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/records/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Records",
        "description": "获取设备的服药记录（供子女端 BFF 调用，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_records_api_v1_public_device_records__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/chat_history/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Chat History",
        "description": "获取设备相关的聊天历史（供子女端 BFF 调用，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_chat_history_api_v1_public_device_chat_history__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/medication_plan/{plan_id}": {
      "delete": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Delete Device Medication Plan",
        "description": "删除用药计划（校验 device_id 与 X-Device-Token 及设备归属）",
        "operationId": "delete_device_medication_plan_api_v1_public_device_medication_plan__plan_id__delete",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Update Device Medication Plan",
        "description": "更新用药计划（校验 device_id 与 X-Device-Token 及设备归属）",
        "operationId": "update_device_medication_plan_api_v1_public_device_medication_plan__plan_id__put",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMedicationPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FamilyMedicationPlan": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "drug_name": {
            "type": "string",
            "title": "Drug Name"
          },
          "dosage": {
            "type": "string",
            "title": "Dosage",
            "default": "1片"
          },
          "product_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Code"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency",
            "default": "每日"
          },
          "schedule_times": {
            "items": {},
            "type": "array",
            "title": "Schedule Times"
          },
          "total_quantity": {
            "type": "number",
            "title": "Total Quantity",
            "default": 30.0
          },
          "remaining_quantity": {
            "type": "number",
            "title": "Remaining Quantity",
            "default": 30.0
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "片"
          },
          "low_stock_threshold": {
            "type": "integer",
            "title": "Low Stock Threshold",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "drug_name",
          "schedule_times"
        ],
        "title": "FamilyMedicationPlan",
        "description": "家属设置用药计划（通过设备ID）"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}